Benchmarking with Processing version performed
Benchmarking results for C++ and Processing
Execution time:
 1. 2,33 | 1,69 - environment and cell rendering
 2. 2,33 | 1,38 - cell rendering
 3. 2,20 | 1,00 - no rendering
Rendering to computation ratio:
 1. 0,06 | 0,69 - environment and cell rendering
 2. 0,06 | 0,38 - cell rendering
 3. 0,00 | 0,31 - environment rendering
C++ version was compiled with -O3 flag
and Processing version was exported
with embedded Java and was using P2D renderer.

It seems rendering part is faster in C++ version
(it is almost instantaneous),
however, computations are much slower than in Java.

Maybe memory allocations (budding) are not
as expensive and/or as frequent as copying (moving),
so it should be better to replace
Cell buffer with buffer of Cell pointers.
It would also be better to replace vectors with arrays.
Although it would cause using templates for Cell
and CellController, it can bring some performance.

Add "-O3" optimization flag.

Update README.md.